refactor(task): split service.go into domain-focused files
Split the 2858-line service.go into four domain-focused files:
- task_service.go: core CRUD, bulk ops, field updates
- template_service.go: recurring/template operations
- flex_service.go: flex task operations
- query_service.go: query/listing operations

Reminder service methods moved to reminder.go. No logic changes.
